Update reviewed items
